backend/push-notify/requirements.txt
psycopg2-binary>=2.9.0
pywebpush>=2.0.0
py-vapid>=1.9.0
